GeoKD-SR/requirements.txt
# ===========================================
# GeoKD-SR 项目依赖
# 适配阿里云PAI平台 (CUDA 12.4)
# 最后更新: 2026年3月1日
# ===========================================

# Deep Learning Framework (PAI已安装)
# torch>=2.6.0          # PAI预装
# torchvision>=0.21.0   # PAI预装
# torchaudio>=2.6.0     # PAI预装
# triton>=3.2.0         # PAI预装

# Transformers & LLM (核心依赖)
transformers>=4.37.0
huggingface-hub>=0.20.0
accelerate>=0.25.0
safetensors>=0.4.0
peft>=0.7.0
datasets>=2.15.0
bitsandbytes>=0.41.0
tqdm>=4.66.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 空间计算 (GeoKD-SR特需)
shapely>=2.0.2
geopy>=2.4.1
pyproj>=3.6.1

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 实验跟踪
wandb>=0.15.0
tensorboard>=2.14.0

# 工具
pyyaml>=6.0
requests>=2.31.0
